requirements.txt
supabase
faker
python-dotenv